Extract ElapsedTimeTracker into its own file
