+++ CodeBreaker/CodeBreaker/UI/CodeBreakerView.swift
@@ -111,43 +111,6 @@ struct CodeBreakerView: View {
     }
 }
 
-extension View {
-    func trackElapsedTime(in game: CodeBreaker) -> some View {
-        modifier(ElapsedTimeTracker(game: game))
-    }
-}
-
-struct ElapsedTimeTracker: ViewModifier {
-    // MARK: Data In
-    @Environment(\.scenePhase) private var scenePhase
-    let game: CodeBreaker
-    
-    func body(content: Content) -> some View {
-        content
-            .onAppear {
-                game.startTimer()
-            }
-            .onDisappear {
-                game.pauseTimer()
-            }
-            .onChange(of: game) { oldGame, newGame in
-                oldGame.pauseTimer()
-                newGame.startTimer()
-            }
-            .onChange(of: scenePhase) {
-                switch scenePhase {
-                case .background:
-                    game.pauseTimer()
-                case .active:
-                    game.startTimer()
-                default:
-                    break
-                }
-            }
-        
-    }
-}
-
 extension CodeBreaker {
     convenience init(name: String = "Code Breaker", pegChoices: [Color]) {
         self.init(name: name, pegChoices: pegChoices.map(\.hex))
